Separate build functions from read/sidebar assembly functions
Reorganizes the page contents utilities to clearly separate concerns:

**page-contents.ts** (Building/Writing):
- Functions for creating and building the Contents structure
- Functions for adding content (addOverviewPage, addToolingPage, etc.)
- writeContents() for writing to JSON
- Used during the generation process to BUILD contents.json

**sidebar-from-contents.ts** (Reading/Assembly):
- readContents() for reading from JSON file
- assembleSidebarFromContents() for creating sidebar from Contents
- Used to READ contents.json and assemble sidebar from it

This separation makes it clear which functions are used to:
1. Generate/build the contents structure (page-contents.ts)
2. Read and use the contents structure (sidebar-from-contents.ts)

No changes to existing build process - all original sidebar building
and JSON file outputs remain unchanged. This is purely organizational.
